+++ packages/primeng/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/ngx-json-form-stepper.component';
 export * from './lib/ngx-json-form-debug.component';
 export * from './lib/signature-pad.component';
 export * from './lib/time-slots.component';
+export * from './lib/markdown-editor.component';